fix: Modify how tRIBS soil parameters are calculated form the ROSETTA3 outputs.
This chnage marks a somewhat major change and bugfix from the previous version. Previously the calculated of the PsiB parameter was incorrectly converted into 1/mm which resulted in the PsiB value being 100 times smaller than it should be. The new PsiB values are not within a range that tRIBS normally sees and will cause the model to produce weird results.

The modification to how the variable, m or the pore distribution index is a better method to estimate m from the ROSETTA3 output. The previous method is for the van Genuchten model which is not applicable to tRIBS.